# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: simone.francalancia@accenture.com
    name: Simone Francalancia
    url: https://github.com/Simo-univpm/demo-app
  description: "Documentazione delle API RESTful esposte dal mio progetto demo svolto\
    \ in Accenture. <br>Queste ruotano attorno ad un utente e ci permettono di registrare,\
    \ loggare, editare <br>un utente e di visualizzarne i dati.<br>Inoltre ci permettono\
    \ di ottenere la data dell' ultimo accesso dell'utente e di generare <br>dei report\
    \ in formato csv riguardanti le varie operazioni effettuate sul server. <br>"
  title: API demo-app
  version: 1.0.0
servers:
- description: Generated server url
  url: http://localhost:8080
paths:
  /api/user:
    put:
      operationId: editUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEditRequest'
        required: true
      responses:
        "404":
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/UserEditRequest'
          description: if the user we want to edit is not registered
        "200":
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/UserEditRequest'
          description: if the specified user is registered and gets correctly edited
      summary: Edit the first_name and last_name fields of the user with the username
        specified in the body
      tags:
      - user-controller
      x-content-type: application/json
      x-accepts: application/hal+json
      x-tags:
      - tag: user-controller
  /api/user/register:
    post:
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
        required: true
      responses:
        "400":
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/UserEntity'
          description: if there is already an user with that specific username
        "200":
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/UserEntity'
          description: if user gets successfully registered
      summary: Register a new user
      tags:
      - user-controller
      x-content-type: application/json
      x-accepts: application/hal+json
      x-tags:
      - tag: user-controller
  /api/user/login:
    post:
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "404":
          description: if the user we want to log-in is not registered
        "200":
          description: if the user gets successfully logged in
        "400":
          description: if the specified user's credentials does not satisfy the application's
            criteria
      summary: Perform log-in with an existing user
      tags:
      - user-controller
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: user-controller
  /api/report:
    post:
      description: The api also returns the number of performed operations
      operationId: writeReport
      parameters:
      - explode: true
        in: query
        name: startDate
        required: true
        schema:
          format: date-time
          type: string
        style: form
      - explode: true
        in: query
        name: endDate
        required: true
        schema:
          format: date-time
          type: string
        style: form
      - explode: true
        in: query
        name: key
        required: true
        schema:
          type: string
        style: form
      responses:
        "400":
          content:
            application/hal+json:
              schema:
                format: int64
                type: integer
          description: if the csv file cannot be opened
        "200":
          content:
            application/hal+json:
              schema:
                format: int64
                type: integer
          description: if the csv file is written correctly
      summary: "Write a csv file with the number of performed operations between the\
        \ specified timestamps. If the file already exists, subsequent operations\
        \ will be appended"
      tags:
      - report-controller
      x-accepts: application/hal+json
      x-tags:
      - tag: report-controller
  /api/report/verbose:
    post:
      operationId: writeReportVerbose
      parameters:
      - explode: true
        in: query
        name: startDate
        required: true
        schema:
          format: date-time
          type: string
        style: form
      - explode: true
        in: query
        name: endDate
        required: true
        schema:
          format: date-time
          type: string
        style: form
      - explode: true
        in: query
        name: key
        required: true
        schema:
          type: string
        style: form
      responses:
        "400":
          description: if the csv file cannot be opened
        "200":
          description: if the csv file is written correctly
      summary: "Write a csv file with the performed operations between the specified\
        \ timestamps. If the file already exists, subsequent operations will be appended"
      tags:
      - report-controller
      x-accepts: application/json
      x-tags:
      - tag: report-controller
  /api/user/{id}:
    delete:
      operationId: deleteUser
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "404":
          description: if the user we want to delete is not registered
        "200":
          description: if the user is successfully deleted
      summary: Deletes the user with the id provided in the id path variable
      tags:
      - user-controller
      x-accepts: application/json
      x-tags:
      - tag: user-controller
    get:
      operationId: getUserDetails
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "404":
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/UserEntity'
          description: if the user we want to view is not registered
        "200":
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/UserEntity'
          description: a list of strings with the data of the requested user
      summary: Return the details of the user specified in the id path variable
      tags:
      - user-controller
      x-accepts: application/hal+json
      x-tags:
      - tag: user-controller
  /api/access/{id}:
    get:
      operationId: getUserLastAccess
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/LastAccessEntity'
          description: if user's last access is returned successfully
        "400":
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/LastAccessEntity'
          description: if user registered but never logged in
      summary: Returns date and time of user's last access to the server
      tags:
      - last-access-controller
      x-accepts: application/hal+json
      x-tags:
      - tag: last-access-controller
components:
  schemas:
    LastAccessEntity:
      properties:
        id:
          format: int64
          type: integer
        userId:
          format: int64
          type: integer
        last_access:
          format: date-time
          type: string
      type: object
    UserEntity:
      properties:
        id:
          format: int64
          type: integer
        username:
          type: string
        passwd:
          type: string
        first_name:
          type: string
        last_name:
          type: string
      type: object
    UserEditRequest:
      example:
        last_name: last_name
        first_name: first_name
        username: username
      properties:
        username:
          maxLength: 20
          minLength: 1
          type: string
        first_name:
          maxLength: 20
          minLength: 1
          type: string
        last_name:
          maxLength: 20
          minLength: 1
          type: string
      required:
      - first_name
      - last_name
      - username
      type: object
    RegistrationRequest:
      example:
        passwd: passwd
        last_name: last_name
        first_name: first_name
        username: username
      properties:
        username:
          maxLength: 20
          minLength: 1
          type: string
        passwd:
          maxLength: 20
          minLength: 8
          type: string
        first_name:
          maxLength: 20
          minLength: 1
          type: string
        last_name:
          maxLength: 20
          minLength: 1
          type: string
      required:
      - first_name
      - last_name
      - passwd
      - username
      type: object
    LoginRequest:
      example:
        passwd: passwd
        username: username
      properties:
        username:
          maxLength: 20
          minLength: 1
          type: string
        passwd:
          maxLength: 20
          minLength: 8
          type: string
      required:
      - passwd
      - username
      type: object